refactor(analysis): break the last import cycle and empty its ledger
adversaryHints.ts called suggestNextTechniques() from adversaryEmulation.ts,
and adversaryEmulation.ts imported the technique-id helpers and both record
shapes back from adversaryHints.ts. Each module needed the other loaded first.

Node tolerates that right up until one of those imports is evaluated at module
scope rather than at call time, at which point one side sees `undefined` — and
the failure surfaces as a data bug (no groups matched, no next techniques) far
from the load-order cause. It was the single entry in the import-cycle ledger.

adversaryTechniques.ts now holds exactly what BOTH sides need: AdversaryGroup,
AdversaryHint, normalizeTechniqueId and baseTechniqueId. The direction is
hints -> emulation -> shared, and hints -> shared. Nothing imports back.

Deliberately NOT a dumping ground: suggestNextTechniques stays in emulation and
rankAdversaryGroups stays in hints. Moving either would turn the new module into
the place cycles form next.

adversaryHints.ts re-exports all four names, so its ten importers — the report
writer, markdown, AI synthesis, prompt blocks, playbook matching, d3fend
mapping, ATT&CK mitigations, false-positive similarity and two test helpers —
compile unchanged. Only the declaration site moved.

The ledger is now `[]`, so scripts/check-imports.mjs has stopped being a
ratchet with one grandfathered entry and become a plain rule: no cycles.

The boundary gate correctly refused the new file until it declared a domain
("an unclassified file is an unanswered design question, not a known debt") —
it joins its two siblings in analysis/intel.

Verified: 0 cycles, typecheck clean, adversary suites 32 passed, the wider
analysis + reports suites 4412 passed, every other ratchet unmoved.
