+++ companion/scripts/import-cycles.json
@@ -1,3 +1,1 @@
-[
-  "analysis/adversaryEmulation.ts -> analysis/adversaryHints.ts"
-]
+[]
